Fixes calculation error which was appeared because of wrong data types. Optimizes Feature content

+++ src/main/java/ch/bfh/bti7535/w2016/algorithm/features/GoodWordsFeature.java
@@ -22,7 +22,7 @@ public class GoodWordsFeature extends AbstractFeature {
 				sentencesQuestionAmount += DocumentUtil.countQuestionSenctences(doc);
 		}
 
-		float result = (sentencesAmount > 0.0001) ? sentencesQuestionAmount / sentencesAmount : 0;
+		double result = (sentencesAmount > 0.0001) ? (double) sentencesQuestionAmount / (double) sentencesAmount : 0;
 		setProbability(classification, result);
 	}
 }
